Update calculator v2.2

+++ myProjects/Calculator/v2_0/calculateResult.cpp
@@ -1,7 +1,7 @@
 #include<iostream>
 
 // Вычисляем полученные значения
-int calculateResult(int x, char op, int y) {
+double calculateResult(double x, char op, double y) {
     switch (op)
     {
     case '+':
