+++ myProjects/Calculator/v2_0/printResult.cpp
@@ -1,6 +1,6 @@
 #include<iostream>
 
 // Вывод результата на экран
-void printResult(int result) {
+void printResult(double result) {
     std::cout << "You result is: " << result << std::endl;
 }
